backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.42.0
openai==1.58.1
pdfplumber==0.11.4
python-docx==1.1.2
lark-oapi==1.4.3
python-multipart==0.0.18
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
